Update Upload implementation
The ViewModel is practically the same, but now it uses the new factory methods on the ViewState class.

The ViewState now is the one responsible for mapping the different high level states to the final values needed by the view to represent the states properly.

The fragment now is significantly simpler and has no logic, it literally simply reflects the current ViewState in its UI. This is intended to be the best part of this implementation. If you look at the code it’s pretty much like data binding.

I’m creating a `setVisible` extension just so I can use method reference when observing some of the state.